requirements.txt
chromadb
sentence-transformers
google-generativeai
python-dotenv
tqdm
Pillow
markdown
fastapi
uvicorn
streamlit
